+++ sigstore/_internal/timestamp.py
@@ -26,6 +26,7 @@ from rfc3161_client import (
     TimeStampResponse,
     decode_timestamp_response,
 )
+from rfc3161_client.base import HashAlgorithm
 
 from rfc3161_client.base import HashAlgorithm
 
@@ -95,8 +96,11 @@ class TimestampAuthorityClient:
         # Build the timestamp request
         try:
             timestamp_request = (
-                TimestampRequestBuilder().hash_algorithm(HashAlgorithm.SHA256).data(
-                    signature).nonce(nonce=True).build()
+                TimestampRequestBuilder()
+                .hash_algorithm(HashAlgorithm.SHA256)
+                .data(signature)
+                .nonce(nonce=True)
+                .build()
             )
         except ValueError as error:
             msg = f"invalid request: {error}"